fixed scrollable issue and refactoring
