+++ components/auth/index.tsx
@@ -1,18 +1,20 @@
-export * from './stars';
-export * from './aurora';
-export * from './screen';
-export * from './card';
-export * from './btn';
-export * from './back-btn';
-export * from './progress-dots';
-export * from './tag-chip';
-export * from './status-bar';
-export * from './bottom-nav';
-export * from './line-chart';
-export * from './bar-chart';
-export * from './auth-shell';
-export * from './input';
-export * from './google-btn';
-export * from './divider';
-export * from './toast';
-export * from './spinner';
+export * from "./aurora";
+export * from "./auth-shell";
+export * from "./back-btn";
+export * from "./bar-chart";
+export * from "./bottom-nav";
+export * from "./btn";
+export * from "./card";
+export * from "./divider";
+export * from "./google-btn";
+export * from "./input";
+export * from "./line-chart";
+export * from "./ob-wrap";
+export * from "./progress-dots";
+export * from "./screen";
+export * from "./spinner";
+export * from "./stars";
+export * from "./status-bar";
+export * from "./tag-chip";
+export * from "./toast";
+
